+++ src/instructlab/log.py
@@ -55,6 +55,7 @@ def stdout_stderr_to_logger(logger, log_file):
 def configure_logging(
     *,
     log_level: str,
+    debug_level: int = 0,
     fmt: str = FORMAT,
 ) -> None:
     """Configure logging framework"""
@@ -68,8 +69,30 @@ def configure_logging(
     stream = logging.StreamHandler()
     stream.setFormatter(CustomFormatter(fmt=fmt))
     root.addHandler(stream)
-    root.setLevel(log_level)
 
-    # Set logging level of OpenAI client and httpx library to ERROR to suppress INFO messages
-    logging.getLogger("openai").setLevel(logging.ERROR)
-    logging.getLogger("httpx").setLevel(logging.ERROR)
+    instructlab_lgr = logging.getLogger("instructlab")
+    openai_lgr = logging.getLogger("openai")
+    httpx_lgr = logging.getLogger("httpx")
+
+    # reset loggers to inherit from root logger
+    for lgr in [instructlab_lgr, openai_lgr, httpx_lgr]:
+        lgr.setLevel(logging.NOTSET)
+
+    if log_level == "DEBUG":
+        if debug_level >= 2:
+            # debug level 2 enables debug logging for everything
+            root.setLevel(logging.DEBUG)
+        else:
+            # debug level 1 enables debugging for instructlab
+            # other loggers are set to INFO (except for openai and httpx)
+            root.setLevel(logging.INFO)
+            instructlab_lgr.setLevel(logging.DEBUG)
+    else:
+        # INFO, WARNING, ERROR, FATAL and higher
+        root.setLevel(log_level)
+
+    if log_level != "DEBUG" or debug_level < 2:
+        # Set logging level of OpenAI client and httpx library to ERROR to
+        # suppress INFO messages unless debug_leve is >= 2.
+        openai_lgr.setLevel(logging.ERROR)
+        httpx_lgr.setLevel(logging.ERROR)